Animation implementation and new scripts
Implements all the new animations for the Ranged Enemy. Attack animation is still a little broken.

The way the bat attacks has been changed. The attack hitbox is tied to the 'attacking' frames of the animation.

The Bat and the Ranged Enemy are a bit buggy at the moment. I updated their prefabs with what I hope to be the final scripts. They're not done yet, so they won't function 100% properly. I'm pushing these changes to save my progress before I go to work.