Some Cerb3 work
Cerb3 is in his arena.

Adds an 'Icicle' object(with prefab) that he shoots. It's faster than the fireball.

Cerb3's shoot timer is the shortest of them all at 7 seconds.

His health is at 900 for now. might increase depending on how strong ashes is.

Also added the ground tag. Not used yet but will be soon.

Like Cerb 1, Cerb 2 and Cerb 3 will eventually be spawned using the trigger prefab(once work on them is done)